requirements.txt
selenium
beautifulsoup4
requests
pandas
python-dotenv
pydantic
firecrawl
